roomcare/requirements.txt
# RoomCare - On-device multimodal AI agent
# Python 3.10+

# Core
torch>=2.1.0
torchvision>=0.16.0
transformers>=5.2.0
accelerate>=0.25.0
peft>=0.10.0
bitsandbytes>=0.41.0

# Vision & media (timm required for Gemma 3n vision tower)
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
timm>=1.0.0

# Audio (on-device)
soundfile>=0.12.0
librosa>=0.10.0
faster-whisper>=1.0.0

# Memory & utils
# SQLite is stdlib

# CLI & runtime
tqdm>=4.65.0
